requirements.txt
aiohttp==3.6.2
altgraph==0.17
aniso8601==8.0.0
appdirs==1.4.3
appnope==0.1.0
asgiref==3.2.5
asn1crypto==1.3.0
astroid==2.3.0
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==19.3.0
backcall==0.1.0
behave==1.2.6
behave-pytest==0.1.1
bleach==3.1.4
boto3==1.12.39
botocore==1.15.39
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.1
clickclick==1.2.2
connexion==2.6.0
coverage==5.1
cPython==0.0.5
cryptography==2.9.2
cycler==0.10.0
Cython==0.29.16
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.1.1
distlib==0.3.0
Django==3.0.4
docutils==0.15.2
docx-utils==0.1.2
entrypoints==0.3
enum34==1.1.10
falcon==2.0.0
fastavro==0.22.13
filelock==3.0.12
Flask==1.1.2
Flask-Injector==0.12.3
flask-restplus==0.12.1
Flask-SQLAlchemy==2.1
Flask-WTF==0.14.3
future==0.18.2
glob2==0.7
gunicorn==19.8.1
httpie==2.0.0
hug==2.6.1
idna==2.9
inflection==0.3.1
injector==0.12.0
ipaddress==1.0.23
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.1
jmespath==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.2
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
lxml==4.5.0
macholib==1.14
Mako==1.1.2
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
mistune==0.8.4
mock==4.0.2
modulegraph==0.18
more-itertools==8.2.0
multidict==4.7.5
multiprocess==0.70.9
nbconvert==5.6.1
nbformat==5.0.5
notebook==6.0.3
numpy==1.18.4
openapi-spec-validator==0.2.8
packaging==20.3
pandocfilters==1.4.2
parse==1.15.0
parse-type==0.5.2
parso==0.6.2
pathos==0.2.5
pbr==5.4.4
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
pox==0.2.7
ppft==1.6.6.1
prometheus-client==0.7.1
prompt-toolkit==3.0.5
ptyprocess==0.6.0
py==1.8.1
py2app==0.21
pycparser==2.20
Pygments==2.6.1
PyHamcrest==2.0.2
pylint==2.4.4
pymongo==3.10.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==5.4.1
python-dateutil==2.8.1
python-dotenv==0.12.0
python-mimeparse==1.6.0
pytz==2019.3
PyYAML==5.3
pyzmq==19.0.0
qtconsole==4.7.2
QtPy==1.9.0
requests==2.23.0
s3transfer==0.3.3
scipy==1.4.1
Send2Trash==1.5.0
six==1.14.0
SQLAlchemy==1.3.16
sqlparse==0.3.1
stevedore==1.32.0
swagger-ui-bundle==0.0.8
terminado==0.8.3
testpath==0.4.4
toml==0.10.0
tornado==6.0.4
traitlets==4.3.3
typing==3.7.4.1
typing-extensions==3.7.4.2
urllib3==1.25.8
virtualenv==20.0.10
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.11.2
WTForms==2.2.1
xattr==0.9.7
yarl==1.4.2
